requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
click==8.1.7
httptools==0.6.1
python-multipart==0.0.6
httpx==0.24.1
pytz==2023.3
supabase==2.0.3
python-dotenv==1.0.0

# Google APIs (optional)
google-api-python-client==2.95.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
